requirements.txt
langchain
rich
openai
click
langchain_core
langchain_openai
pyfiglet
python-dotenv